:recycle: refactor : gateway 에서 사용되는 dto 별도 파일로 분리 #75
