+++ backend/src/user/user.gateway.ts
@@ -3,7 +3,7 @@ import { WebSocketGateway, WebSocketServer } from '@nestjs/websockets';
 
 import { SocketId } from '../util/type';
 import { UserId } from '../util/type';
-import { UserInfoDto } from './dto/user.dto';
+import { UserInfoDto } from './dto/user-gateway.dto';
 
 @WebSocketGateway()
 export class UserGateway {
